// sdk/base/lib/osBindings/Progress.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FullProgress } from './FullProgress'
import type { ProgressUnits } from './ProgressUnits'

export type Progress =
  | null
  | boolean
  | { done: number; total: number | null; units: ProgressUnits | null }
  | FullProgress